make an ast package
