+++ pym/render/render.py
@@ -2,9 +2,9 @@
 
 import ast
 
-from .visitors import Renderer
-from ..transform.commenter import add_comments
-from ..transform.docstringer import recast_docstrings
+from .renderer import Renderer
+from ..ast.transform.commenter import add_comments
+from ..ast.transform.docstringer import recast_docstrings
 
 
 def parse(source, path=None):
